requirements.txt
aiogram==3.13.1
replicate==0.34.1
python-dotenv==1.0.1
aiohttp==3.10.11
